// src/modules/weapons/model/record.interface.ts
export interface Record {
    IDAccount: number;
    IDActaTraumatica: number;
    IDFuncionario: number;
    IDOffice: number;
    IDSalidaAlmacen: number;
    Media: string;
    MedicaContext: string;
    Since: string;
    TipoActa: number;
    ListaDoc: string;
    document: string;
    record: string;
    URLdocument: string;
    URLrecord: string;
}